backend /requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.2.3
SQLAlchemy==1.4.46